Update ObjectMembers.class.php
HI there, the line 58 and 53 had  the name of the table foc_object_members hard code, provoking  that anything could be saved,  i changed to use the TABLE_PREFIX variable.

Hopefully this can help, cause i've having issues for mor than a month, moving to an older version. 
Hopping from now on everything is as perfect as in previos versions.